requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
elevenlabs>=1.0.0
openai>=1.0.0
moviepy>=2.0.0
numpy>=1.21.0
Pillow>=10.0.0
scipy>=1.7.0
pyyaml>=6.0
requests>=2.28.0
